# src/swagger.json
{
  "swagger": "2.0",
  "info": {
    "description": "",
    "version": "1.0",
    "title": "Dasa Challenge - Back"
  },
  "host": "localhost:3030",
...
  "schemes": ["http"],
  "paths": {
...
    "/laboratory/destroy/{labsIds}": {
      "delete": {
        "tags": ["Laboratory"],
        "summary": "Delete one or more laboratories in the repository",
        "description": "",
        "operationId": "destroy",
        "produces": ["application/json"],
        "parameters": [
          {
            "name": "labsIds",
            "in": "path",
            "description": "String with laboratories IDs (separete by '|') that needs to be deleted at the repository",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "404": {
            "description": "Action canceled! Laboratory not found at positions:",
            "schema": {
              "$ref": "#/definitions/Error"
            }
          }
        }
      }
    },
...
    "/exam/destroy/{examsIds}": {
      "delete": {
        "tags": ["Exam"],
        "summary": "Delete one or more exams in the repository",
        "description": "",
        "operationId": "destroy",
        "produces": ["application/json"],
        "parameters": [
          {
            "name": "examsIds",
            "in": "path",
            "description": "String with exams IDs (separete by '|') that needs to be deleted at the repository",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "404": {
            "description": "Action canceled! Exam not found at positions:",
            "schema": {
              "$ref": "#/definitions/Error"
            }
          }
        }
      }
    },
    "/exam/{examId}/linkLaboratory": {
      "post": {
        "tags": ["Exam"],
        "summary": "Link one or more laboratories to a specific exam",
        "description": "",
        "operationId": "linkLaboratory",
        "consumes": ["application/json"],
        "produces": ["application/json"],
        "parameters": [
          {
            "name": "examId",
            "in": "path",
            "description": "Exam identification code",
            "required": true,
            "type": "string"
          },
          {
            "in": "body",
            "name": "body",
            "description": "Array of id laboratory that needs to be link",
            "required": true,
            "schema": { "$ref": "#/definitions/ArrayOfIdLaboratories" }
          }
        ],
        "responses": {
          "404": {
            "description": "Action canceled! Exam not found",
            "schema": {
              "$ref": "#/definitions/Error"
            }
          },
          "405": {
            "description": "Action canceled! Invalid input at positions:",
            "schema": {
              "$ref": "#/definitions/Error"
            }
          }
        }
      }
    },
    "/exam/{examId}/unlinkLaboratory": {
      "delete": {
        "tags": ["Exam"],
        "summary": "Unlink one or more laboratories to a specific exam",
        "description": "",
        "operationId": "unlinkLaboratory",
        "consumes": ["application/json"],
        "produces": ["application/json"],
        "parameters": [
          {
            "name": "examId",
            "in": "path",
            "description": "Exam identification code",
            "required": true,
            "type": "string"
          },
          {
            "in": "body",
            "name": "body",
            "description": "Array of id laboratory that needs to be unlink",
            "required": true,
            "schema": { "$ref": "#/definitions/ArrayOfIdLaboratories" }
          }
        ],
        "responses": {
          "404": {
            "description": "Action canceled! Exam not found",
            "schema": {
              "$ref": "#/definitions/Error"
            }
          },
          "405": {
            "description": "Action canceled! Invalid input at positions:",
            "schema": {
              "$ref": "#/definitions/Error"
            }
          }
        }
      }
    },
...
  },
  "definitions": {
...
    "ArrayOfIdLaboratories": {
      "type": "array",
      "items": {
        "type": "string"
      }
    },
...
    "Error": {
      "type": "object",
      "properties": {
        "message": { "type": "string" },
        "positions": {
          "type": "array",
          "items": {
            "type": "integer",
            "format": "int32"
          }
        }
      }
    }
  }
}